requirements.txt
numpy==1.23.2
scikit-learn==1.1.3
spacy==3.3
flask==2.0.1
tensorflow==2.10.0
tensorflow-text==2.10.0
tensorflow-hub==0.10.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.3.0/en_core_web_lg-3.3.0.tar.gz
waitress==2.1.2
mysql_connector==2.2.9
faiss_gpu==1.7.2
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
sklearn==0.0.post1
google-cloud-storage==2.5.0
PyYAML==6.0
pathlib==1.0.1
pandas==1.4.4
